+++ FinalProject/WebContent/WEB-INF/view/LoginForm.jsp
@@ -59,20 +59,3 @@
 		</form>
 	</div>
 </div>
-<script>
-	function check()
-	{
-
-		if ($.trim($("#user_id").val()) == "")
-		{
-			alert("아이디를 입력해주세요");
-			return false;
-		}
-		if ($.trim($("#user_pw").val()) == "")
-		{
-			alert("비밀번호를 입력해주세요");
-			return false;
-		}
-
-	}
-</script>
