--- FinalProject/WebContent/WEB-INF/view/RegionListAjax.jsp
@@ -1,11 +1,0 @@
-<%@ page contentType="text/html; charset=UTF-8"%>
-<%@ taglib prefix="c" uri="http://java.sun.com/jsp/jstl/core"%>
-<% 
-	request.setCharacterEncoding("UTF-8");
-	String cp = request.getContextPath();
-%>
-<datalist id="region">
-<c:forEach var="region" items="${regionList }">
-<option value="${region.region_name}"></option>
-</c:forEach>
-</datalist>
