--- FinalProject/src/com/kkini/dto/RegionDTO.java
@@ -1,29 +1,0 @@
-package com.kkini.dto;
-
-public class RegionDTO
-{
-	String region_name, region_code;
-
-	public String getRegion_name()
-	{
-		return region_name;
-	}
-
-	public void setRegion_name(String region_name)
-	{
-		this.region_name = region_name;
-	}
-
-	public String getRegion_code()
-	{
-		return region_code;
-	}
-
-	public void setRegion_code(String region_code)
-	{
-		this.region_code = region_code;
-	}
-
-	
-	
-}
