+++ FinalProject/src/com/kkini/dto/UserDTO.java
@@ -6,10 +6,31 @@ public class UserDTO
 	String user_id, user_pw, user_nickname, user_birth, user_gender, user_addr, user_phonenumber;
 	String[] user_intregions, user_interests, user_characters;
 	String user_eatSpeed, user_talk, user_introduce;
-	String attendScore, mannerScore, biasScore;
+	String attendScore, mannerScore, biasScore, region_name, region_code;
 	
 	String user_code;
 	String user_code_live;
+	
+	
+	
+	
+	
+	public String getRegion_name()
+	{
+		return region_name;
+	}
+	public void setRegion_name(String region_name)
+	{
+		this.region_name = region_name;
+	}
+	public String getRegion_code()
+	{
+		return region_code;
+	}
+	public void setRegion_code(String region_code)
+	{
+		this.region_code = region_code;
+	}
 	public String getUser_id()
 	{
 		return user_id;
