--- FinalProject/src/com/kkini/mybatis/IRegionDAO.java
@@ -1,14 +1,0 @@
-package com.kkini.mybatis;
-
-import java.util.ArrayList;
-
-import com.kkini.dto.RegionDTO;
-
-public interface IRegionDAO
-{
-	
-	public ArrayList<RegionDTO> list(String region_name);
-	
-	public int check(String region_name);
-	
-}
