--- FinalProject/src/com/kkini/mybatis/mapper/RegionDAO.xml
@@ -1,23 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org/DTD Mapper 3.0/EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-
-<mapper namespace="com.kkini.mybatis.IRegionDAO">
-
-<select id="list" resultType="com.kkini.dto.RegionDTO" parameterType="java.lang.String">
-SELECT REGION_NAME, REGION_CODE
-FROM REGION
-WHERE REGION_NAME LIKE '%' || #{region_name} || '%'
-</select>
-
-
-<select id="check" resultType="java.lang.Integer" parameterType="java.lang.String">
-SELECT COUNT(*) AS COUNT
-FROM REGION
-WHERE REGION_NAME LIKE #{region_name}
-
-</select>
-
-
-</mapper>
